requirements.txt
# Core dependencies
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<3.0.0

# Data collection
yfinance>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Code quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0

# Development
ipython>=8.12.0
jupyter>=1.0.0
notebook>=6.5.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0

# Logging & Utils
loguru>=0.6.0
tqdm>=4.65.0
